Moved types file
