+++ packages/nextjs/src/types.ts
